requirements.txt
django
requests
django-cors-headers
djangorestframework
PyJWT
pytz